revert only installing Network Location in system user
We need to be installed in non-main (secondary) users to work in them users due to an upstream bug.
https://github.com/GrapheneOS/platform_frameworks_base/blob/dc379a48d1e7b456649d04daa66bbf00a3f32e2c/services/core/java/com/android/server/servicewatcher/CurrentUserServiceSupplier.java#L308
"It queries service as current user, then overrides userId to 0 (main/system user) if serviceIsMultiuser is true.
So non-userId-0 service should never be started but it needs to be installed anyway."
explanation from @muhomorr