make best result selection not take Z accuracy radius into account
This was causing issues due to the fact that it doesn't allow results with a better
XY accuracy radius but null altitude if the best result had an altitude.

Having accurate altitude for measurements should always result in better XZ accuracy
radius for trilateration vs without altitude for the same measurements, so there's
no point in comparing altitude accuracy too and only complicates things.